refactoring inference
