--- examples/inference/src/slt_inference/util.py
@@ -1,265 +1,0 @@
-from pathlib import Path
-from typing import Sequence
-
-import torch
-
-SIGNHIERA_EXPECTED_MISSING = {
-    "norm.weight",
-    "norm.bias",
-    "head.projection.weight",
-    "head.projection.bias",
-}
-SIGNHIERA_EXPECTED_EXTRA = ("decoder", "multi_scale_fusion", "encoder_norm", "mask_token")
-
-
-def load_model(model: torch.nn.Module, checkpoint_path: Path, model_key: str = "model") -> None:
-    """
-    Loads only the model from a saved checkpoint
-    Model loading is not strict and parameters are evicted from the loaded state_dict if their
-    shape does not match the one in passed `model`.
-    """
-
-    with open(checkpoint_path, "rb") as f:
-        checkpoint = torch.load(f, map_location="cpu")
-
-    # Try all of these if necessary
-    for candidate_key in [model_key, "model", "model_state"]:
-        if candidate_key in checkpoint.keys():
-            checkpoint_model = checkpoint[candidate_key]
-            break
-
-    new_checkpoint_model = {}
-    for k, v in checkpoint_model.items():
-        if "feature_proj" in k and "feature_projection" not in k:
-            k = k.replace("feature_proj", "feature_projection.feature_proj")
-
-        if k in model.state_dict() and model.state_dict()[k].shape != v.shape:
-            print(f"Pruning {k} due to size mismatch")
-        else:
-            new_checkpoint_model[k] = v
-
-    missing_keys, unexpected_keys = model.load_state_dict(new_checkpoint_model, strict=False)
-    # Filter out keys expected to be missing or extra when loading SignHiera
-    missing_keys = [k for k in missing_keys if k not in SIGNHIERA_EXPECTED_MISSING]
-    unexpected_keys = [k for k in unexpected_keys if not k.startswith(SIGNHIERA_EXPECTED_EXTRA)]
-    if not (missing_keys or unexpected_keys):
-        print("All keys matched successfully\n")
-    else:
-        print(f"{missing_keys = }, {unexpected_keys = }\n")
-
-
-def print_translations(keys: Sequence[str], translations: Sequence[str]) -> None:
-    assert len(keys) == len(translations)
-    print(f"\nTranslations:\n{'-'*50}")
-    for key, translation in zip(keys, translations):
-        print(f'{key}: "{translation}"')
-    print(f"{'-' * 50}\n")
-
-
-FLORES200_LANG2ID = {
-    "Acehnese (Arabic script)": "ace_Arab",
-    "Acehnese (Latin script)": "ace_Latn",
-    "Mesopotamian Arabic": "acm_Arab",
-    "Ta’izzi-Adeni Arabic": "acq_Arab",
-    "Tunisian Arabic": "aeb_Arab",
-    "Afrikaans": "afr_Latn",
-    "South Levantine Arabic": "ajp_Arab",
-    "Akan": "aka_Latn",
-    "Amharic": "amh_Ethi",
-    "North Levantine Arabic": "apc_Arab",
-    "Modern Standard Arabic": "arb_Arab",
-    "Modern Standard Arabic (Romanized)": "arb_Latn",
-    "Najdi Arabic": "ars_Arab",
-    "Moroccan Arabic": "ary_Arab",
-    "Egyptian Arabic": "arz_Arab",
-    "Assamese": "asm_Beng",
-    "Asturian": "ast_Latn",
-    "Awadhi": "awa_Deva",
-    "Central Aymara": "ayr_Latn",
-    "South Azerbaijani": "azb_Arab",
-    "North Azerbaijani": "azj_Latn",
-    "Bashkir": "bak_Cyrl",
-    "Bambara": "bam_Latn",
-    "Balinese": "ban_Latn",
-    "Belarusian": "bel_Cyrl",
-    "Bemba": "bem_Latn",
-    "Bengali": "ben_Beng",
-    "Bhojpuri": "bho_Deva",
-    "Banjar (Arabic script)": "bjn_Arab",
-    "Banjar (Latin script)": "bjn_Latn",
-    "Standard Tibetan": "bod_Tibt",
-    "Bosnian": "bos_Latn",
-    "Buginese": "bug_Latn",
-    "Bulgarian": "bul_Cyrl",
-    "Catalan": "cat_Latn",
-    "Cebuano": "ceb_Latn",
-    "Czech": "ces_Latn",
-    "Chokwe": "cjk_Latn",
-    "Central Kurdish": "ckb_Arab",
-    "Crimean Tatar": "crh_Latn",
-    "Welsh": "cym_Latn",
-    "Danish": "dan_Latn",
-    "German": "deu_Latn",
-    "Southwestern Dinka": "dik_Latn",
-    "Dyula": "dyu_Latn",
-    "Dzongkha": "dzo_Tibt",
-    "Greek": "ell_Grek",
-    "English": "eng_Latn",
-    "Esperanto": "epo_Latn",
-    "Estonian": "est_Latn",
-    "Basque": "eus_Latn",
-    "Ewe": "ewe_Latn",
-    "Faroese": "fao_Latn",
-    "Fijian": "fij_Latn",
-    "Finnish": "fin_Latn",
-    "Fon": "fon_Latn",
-    "French": "fra_Latn",
-    "Friulian": "fur_Latn",
-    "Nigerian Fulfulde": "fuv_Latn",
-    "Scottish Gaelic": "gla_Latn",
-    "Irish": "gle_Latn",
-    "Galician": "glg_Latn",
-    "Guarani": "grn_Latn",
-    "Gujarati": "guj_Gujr",
-    "Haitian Creole": "hat_Latn",
-    "Hausa": "hau_Latn",
-    "Hebrew": "heb_Hebr",
-    "Hindi": "hin_Deva",
-    "Chhattisgarhi": "hne_Deva",
-    "Croatian": "hrv_Latn",
-    "Hungarian": "hun_Latn",
-    "Armenian": "hye_Armn",
-    "Igbo": "ibo_Latn",
-    "Ilocano": "ilo_Latn",
-    "Indonesian": "ind_Latn",
-    "Icelandic": "isl_Latn",
-    "Italian": "ita_Latn",
-    "Javanese": "jav_Latn",
-    "Japanese": "jpn_Jpan",
-    "Kabyle": "kab_Latn",
-    "Jingpho": "kac_Latn",
-    "Kamba": "kam_Latn",
-    "Kannada": "kan_Knda",
-    "Kashmiri (Arabic script)": "kas_Arab",
-    "Kashmiri (Devanagari script)": "kas_Deva",
-    "Georgian": "kat_Geor",
-    "Central Kanuri (Arabic script)": "knc_Arab",
-    "Central Kanuri (Latin script)": "knc_Latn",
-    "Kazakh": "kaz_Cyrl",
-    "Kabiyè": "kbp_Latn",
-    "Kabuverdianu": "kea_Latn",
-    "Khmer": "khm_Khmr",
-    "Kikuyu": "kik_Latn",
-    "Kinyarwanda": "kin_Latn",
-    "Kyrgyz": "kir_Cyrl",
-    "Kimbundu": "kmb_Latn",
-    "Northern Kurdish": "kmr_Latn",
-    "Kikongo": "kon_Latn",
-    "Korean": "kor_Hang",
-    "Lao": "lao_Laoo",
-    "Ligurian": "lij_Latn",
-    "Limburgish": "lim_Latn",
-    "Lingala": "lin_Latn",
-    "Lithuanian": "lit_Latn",
-    "Lombard": "lmo_Latn",
-    "Latgalian": "ltg_Latn",
-    "Luxembourgish": "ltz_Latn",
-    "Luba-Kasai": "lua_Latn",
-    "Ganda": "lug_Latn",
-    "Luo": "luo_Latn",
-    "Mizo": "lus_Latn",
-    "Standard Latvian": "lvs_Latn",
-    "Magahi": "mag_Deva",
-    "Maithili": "mai_Deva",
-    "Malayalam": "mal_Mlym",
-    "Marathi": "mar_Deva",
-    "Minangkabau (Arabic script)": "min_Arab",
-    "Minangkabau (Latin script)": "min_Latn",
-    "Macedonian": "mkd_Cyrl",
-    "Plateau Malagasy": "plt_Latn",
-    "Maltese": "mlt_Latn",
-    "Meitei (Bengali script)": "mni_Beng",
-    "Halh Mongolian": "khk_Cyrl",
-    "Mossi": "mos_Latn",
-    "Maori": "mri_Latn",
-    "Burmese": "mya_Mymr",
-    "Dutch": "nld_Latn",
-    "Norwegian Nynorsk": "nno_Latn",
-    "Norwegian Bokmål": "nob_Latn",
-    "Nepali": "npi_Deva",
-    "Northern Sotho": "nso_Latn",
-    "Nuer": "nus_Latn",
-    "Nyanja": "nya_Latn",
-    "Occitan": "oci_Latn",
-    "West Central Oromo": "gaz_Latn",
-    "Odia": "ory_Orya",
-    "Pangasinan": "pag_Latn",
-    "Eastern Panjabi": "pan_Guru",
-    "Papiamento": "pap_Latn",
-    "Western Persian": "pes_Arab",
-    "Polish": "pol_Latn",
-    "Portuguese": "por_Latn",
-    "Dari": "prs_Arab",
-    "Southern Pashto": "pbt_Arab",
-    "Ayacucho Quechua": "quy_Latn",
-    "Romanian": "ron_Latn",
-    "Rundi": "run_Latn",
-    "Russian": "rus_Cyrl",
-    "Sango": "sag_Latn",
-    "Sanskrit": "san_Deva",
-    "Santali": "sat_Olck",
-    "Sicilian": "scn_Latn",
-    "Shan": "shn_Mymr",
-    "Sinhala": "sin_Sinh",
-    "Slovak": "slk_Latn",
-    "Slovenian": "slv_Latn",
-    "Samoan": "smo_Latn",
-    "Shona": "sna_Latn",
-    "Sindhi": "snd_Arab",
-    "Somali": "som_Latn",
-    "Southern Sotho": "sot_Latn",
-    "Spanish": "spa_Latn",
-    "Tosk Albanian": "als_Latn",
-    "Sardinian": "srd_Latn",
-    "Serbian": "srp_Cyrl",
-    "Swati": "ssw_Latn",
-    "Sundanese": "sun_Latn",
-    "Swedish": "swe_Latn",
-    "Swahili": "swh_Latn",
-    "Silesian": "szl_Latn",
-    "Tamil": "tam_Taml",
-    "Tatar": "tat_Cyrl",
-    "Telugu": "tel_Telu",
-    "Tajik": "tgk_Cyrl",
-    "Tagalog": "tgl_Latn",
-    "Thai": "tha_Thai",
-    "Tigrinya": "tir_Ethi",
-    "Tamasheq (Latin script)": "taq_Latn",
-    "Tamasheq (Tifinagh script)": "taq_Tfng",
-    "Tok Pisin": "tpi_Latn",
-    "Tswana": "tsn_Latn",
-    "Tsonga": "tso_Latn",
-    "Turkmen": "tuk_Latn",
-    "Tumbuka": "tum_Latn",
-    "Turkish": "tur_Latn",
-    "Twi": "twi_Latn",
-    "Central Atlas Tamazight": "tzm_Tfng",
-    "Uyghur": "uig_Arab",
-    "Ukrainian": "ukr_Cyrl",
-    "Umbundu": "umb_Latn",
-    "Urdu": "urd_Arab",
-    "Northern Uzbek": "uzn_Latn",
-    "Venetian": "vec_Latn",
-    "Vietnamese": "vie_Latn",
-    "Waray": "war_Latn",
-    "Wolof": "wol_Latn",
-    "Xhosa": "xho_Latn",
-    "Eastern Yiddish": "ydd_Hebr",
-    "Yoruba": "yor_Latn",
-    "Yue Chinese": "yue_Hant",
-    "Chinese (Simplified)": "zho_Hans",
-    "Chinese (Traditional)": "zho_Hant",
-    "Standard Malay": "zsm_Latn",
-    "Zulu": "zul_Latn",
-}
-FLORES200_ID2LANG = {v: k for k, v in FLORES200_LANG2ID.items()}
